requirements.txt
jinja2>=1.7
python-daemon==1.6.1
Sphinx>=1.2
numpydoc>=0.4
pandas>=0.14
nose>=1.3
PyYAML>=3.11
psutil>=2.0
enum34>=1.0.4
Flask>=0.10.1
